flask-frontend/pyproject.toml
[project]
name = "flask-frontend"
version = "0.1.0"
description = "Flask frontend for student enrollment system"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "flask>=3.0.0",
    "flask-session>=0.5.0",
    "flask-wtf>=1.2.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "email-validator>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app_package"]